Now display of audio, video, text should be fixed…fingers crossed

+++ honoradar/mysite/honoradar/models.py
@@ -36,11 +36,11 @@ class Medium(models.Model):
 class DataCollection(models.Model):
 
     Medium = models.ForeignKey(Medium, on_delete=models.CASCADE)
-    Happiness=models.IntegerField(default=0)
-    SalaryPerMonthEmpMix=models.IntegerField(default=0)
-    FeeFree=models.IntegerField(default=0)
-    SalaryPerHour=models.IntegerField(default=0)
-    SalaryPerMonth=models.IntegerField(default=0)
+    Happiness=models.FloatField(default=0)
+    SalaryPerMonthEmpMix=models.FloatField(default=0)
+    FeeFree=models.FloatField(default=0)
+    SalaryPerHour=models.FloatField(default=0)
+    SalaryPerMonth=models.FloatField(default=0)
 
 
     #Have to change into buttons
@@ -48,11 +48,11 @@ class DataCollection(models.Model):
     EXPERIENCE = Choices("keineAngabe","1 Jahr","3 Jahre"," 5 Jahre")
     Experience = models.CharField(choices=EXPERIENCE, default=EXPERIENCE.keineAngabe, max_length=10, null=True)
 
-    HoursPerWeekEmp=models.IntegerField(default=0)
-    HoursSpentFree=models.IntegerField(default=0)
+    HoursPerWeekEmp=models.FloatField(default=0)
+    HoursSpentFree=models.FloatField(default=0)
 
-    HoursPerDayMix=models.IntegerField(default=0)
-    DaysPerMonthMix=models.IntegerField(default=0)
+    HoursPerDayMix=models.FloatField(default=0)
+    DaysPerMonthMix=models.FloatField(default=0)
     Genre=models.CharField(default="None",max_length=200)
 
     VATF = Choices("keineAngabe","text","audio"," video")
@@ -62,9 +62,9 @@ class DataCollection(models.Model):
     AnalogDigitalFree = models.CharField(choices=ANALOGDIGITAL, default=ANALOGDIGITAL.keineAngabe, max_length=10, null=True)
 
 
-    MinPerAudioFree=models.IntegerField(default=0)
-    MinPerVideoFree=models.IntegerField(default=0)
-    CharPerArticleFree=models.IntegerField(default=0)
+    MinPerAudioFree=models.FloatField(default=0)
+    MinPerVideoFree=models.FloatField(default=0)
+    CharPerArticleFree=models.FloatField(default=0)
 
     Comment=models.CharField(default="Kein Kommentar",max_length=600, null=True)
